Fix error during import
I haven't looked into why exactly this happened during an import and not any other time, but I got this error:

Resource data.aws_region.current does not have either "count" or "for_each" set, so it cannot be indexed.

It kind of makes sense. This local variable will break if the module is not enabled, because it won't always have a [0] to access. The error message doesn't really make sense though.