Fix STS AssumeRoles
Allowing STS AssumeRole on '*' is incredibly scary. This basically would
allow someone that gains access to an instance to assume any role that
had an assume role policy that allows EC2 to assume it. This means if
they gain enough access to communicate with http://169.254.169.254/ to
get access to the instance role endpoint they basically gain the ability
to become any of the other instance types. In particular I would worry
about them assuming a Jenkins instance's role because these roles accrue
much access including the ability to update Lambda functions, S3 buckets
and in some cases to run Cloudformation. Once they have that level of
access they could completely decimate the account's infrastructure.

Changing the assumerole permission to require a list of IAM role ARNs to
allow the instance to assume. This is intentionally a breaking change
and we should aim to update all instances using this policy ASAP.