Remove ssmparameter_allowall
This policy is so rarely appropriate that it probably shouldn't be an option. If anyone really needs this functionality, please use a better variable name and don't use * by default. Make the user explicitly pass in * if they need the instance to have access to all parameters in the account.